+++ changelog.md
@@ -32,6 +32,8 @@ Codion Change Log
 - EntityTableExportPanel, added save and open configuration.
 - EntityTablePanel.ColumnSelection, AutoResizeModeSelection and EditAttributeSelection enums combined into SelectionMode.
 - EntityTablePanel.Config.includeExport() added along with INCLUDE_EXPORT configuration property, default false.
+- EntityTablePanel, copy expanded rebranded as export.
+- EntityTableExportPanel now uses export instead of copy.
 
 ## 0.18.51
 ### is.codion.swing.common.ui